+++ scripts/generateAddresses.ts
@@ -24,7 +24,7 @@ import { gnosisProtoV3 } from 'scripts/configs/pools/gnosis';
 import { bnbProtoV3 } from 'scripts/configs/pools/bnb';
 import { optimismProtoV3, optimismSepoliaProtoV3 } from 'scripts/configs/pools/optimism';
 import { polygonProtoV2, polygonProtoV3 } from 'scripts/configs/pools/polygon';
-import { scrollSepoliaProtoV3, scrollProtoV3 } from 'scripts/configs/pools/scroll';
+import { scrollProtoV3 } from 'scripts/configs/pools/scroll';
 import { zkSyncProtoV3 } from 'scripts/configs/pools/zksync';
 import { lineaProtoV3 } from 'scripts/configs/pools/linea';
 import { celoProtoV3 } from 'scripts/configs/pools/celo';
@@ -37,6 +37,7 @@ import { megaEthProtoV3 } from 'scripts/configs/pools/megaeth';
 import { xLayerProtoV3 } from 'scripts/configs/pools/xlayer';
 import { monadProtoV3 } from 'scripts/configs/pools/monad';
 import { mainnetV4Config } from 'scripts/configs/v4/ethereum';
+import { avalancheV4Config } from 'scripts/configs/v4/avalanche';
 import { generateProtocolV4Library } from 'scripts/generator/protocolV4Generator';
 import { generateGovernanceLibrary } from 'scripts/generator/governanceV3Generator';
 import { generateProtocolV2Library } from 'scripts/generator/protocolV2Generator';
@@ -180,7 +181,6 @@ async function main() {
       optimismProtoV3,
       optimismSepoliaProtoV3,
       scrollProtoV3,
-      scrollSepoliaProtoV3,
       zkSyncProtoV3,
       // fantomTestnetProtoV3,
       // fantomProtoV3,
@@ -202,7 +202,7 @@ async function main() {
     ].map((config) => generateProtocolV3Library(config)),
   );
   const v4LibraryNames = await Promise.all(
-    [mainnetV4Config].map((config) => generateProtocolV4Library(config)),
+    [mainnetV4Config, avalancheV4Config].map((config) => generateProtocolV4Library(config)),
   );
 
   const ghoAddresses = [ghoEthereum, ghoArbitrum, ghoBase, ghoAvalanche, ghoGnosis, ghoInk, ghoPlasma, ghoMantle, ghoMonad, ghoXLayer].map((config) =>